Selected Penalty

Precomputed output of the penalty selection process where the full output is returned in addition to the penalty selected.

Usage

data(selected_penalty)

Format

A list containing a data frame and numeric.

Details

This object consists of a list with two items:

A data frame containing the full penalty results including the global means of each penalty tested.

The value of the selected penalty

See also